refactor: extract chat response policies and persistence
